Split oversized panel and state test modules
