+++ src/state/mod.rs
@@ -484,147 +484,4 @@ fn write_atomic(path: &Path, contents: &str) -> Result<()> {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn reentrant_lock_succeeds() {
-        let dir =
-            std::env::temp_dir().join(format!("loom-reentrant-{}", uuid::Uuid::new_v4().simple()));
-        let ctx = AppContext::new(Some(dir.clone())).unwrap();
-        let guard1 = ctx.lock_workspace().expect("first lock must succeed");
-        let guard2 = ctx
-            .lock_workspace()
-            .expect("second reentrant lock must succeed");
-        let lock_path = ctx.locks_dir.join("workspace.lock");
-        assert!(
-            lock_path.exists(),
-            "lock file must exist while guards are held"
-        );
-        drop(guard1);
-        drop(guard2);
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn inner_drop_does_not_release_file() {
-        let dir =
-            std::env::temp_dir().join(format!("loom-inner-drop-{}", uuid::Uuid::new_v4().simple()));
-        let ctx = AppContext::new(Some(dir.clone())).unwrap();
-        let guard1 = ctx.lock_workspace().unwrap();
-        let guard2 = ctx.lock_workspace().unwrap();
-        let lock_path = ctx.locks_dir.join("workspace.lock");
-        drop(guard2);
-        assert!(
-            lock_path.exists(),
-            "lock file must exist after inner guard drop"
-        );
-        drop(guard1);
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn outer_drop_releases_file() {
-        let dir =
-            std::env::temp_dir().join(format!("loom-outer-drop-{}", uuid::Uuid::new_v4().simple()));
-        let ctx = AppContext::new(Some(dir.clone())).unwrap();
-        let guard1 = ctx.lock_workspace().unwrap();
-        let guard2 = ctx.lock_workspace().unwrap();
-        let lock_path = ctx.locks_dir.join("workspace.lock");
-        drop(guard2);
-        drop(guard1);
-        assert!(
-            !lock_path.exists(),
-            "lock file must not exist after all guards dropped"
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn cross_context_lock_is_busy() {
-        let dir =
-            std::env::temp_dir().join(format!("loom-cross-ctx-{}", uuid::Uuid::new_v4().simple()));
-        let ctx_a = AppContext::new(Some(dir.clone())).unwrap();
-        let ctx_b = AppContext::new(Some(dir.clone())).unwrap();
-        let _guard = ctx_a.lock_workspace().expect("context A must acquire lock");
-        let result = ctx_b.lock_workspace();
-        assert!(
-            result.is_err(),
-            "context B must not acquire lock while A holds it"
-        );
-        let err_msg = result.unwrap_err().to_string();
-        assert!(
-            err_msg.contains("LOCK_BUSY"),
-            "error must indicate LOCK_BUSY, got: {}",
-            err_msg
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    #[test]
-    fn cloned_context_on_different_thread_is_busy() {
-        // A cloned AppContext shares the same Arc<in_proc> as the original.
-        // A thread holding the lock must block another thread even when that
-        // other thread uses a clone of the same AppContext (the panel path).
-        let dir = std::env::temp_dir().join(format!(
-            "loom-clone-thread-{}",
-            uuid::Uuid::new_v4().simple()
-        ));
-        let ctx_a = AppContext::new(Some(dir.clone())).unwrap();
-        let ctx_b = ctx_a.clone();
-        let _guard = ctx_a
-            .lock_workspace()
-            .expect("main thread must acquire lock");
-        let result = std::thread::spawn(move || ctx_b.lock_workspace())
-            .join()
-            .expect("thread must not panic");
-        assert!(
-            result.is_err(),
-            "cloned context on a different thread must not reenter held lock"
-        );
-        let err_msg = result.unwrap_err().to_string();
-        assert!(
-            err_msg.contains("LOCK_BUSY"),
-            "error must indicate LOCK_BUSY, got: {}",
-            err_msg
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-
-    /// Regression: a cloned AppContext on a different thread must not bypass
-    /// mutual exclusion via stale-lock reaping when the in-proc holder is live.
-    /// Simulates the scenario where the lock file could appear stale (we
-    /// manipulate in_proc directly to verify the guard fires before reaping).
-    #[test]
-    fn cloned_context_different_thread_not_reaped_as_stale() {
-        let dir = std::env::temp_dir().join(format!(
-            "loom-stale-guard-{}",
-            uuid::Uuid::new_v4().simple()
-        ));
-        let ctx_a = AppContext::new(Some(dir.clone())).unwrap();
-        let ctx_b = ctx_a.clone();
-        let _guard = ctx_a
-            .lock_workspace()
-            .expect("main thread must acquire lock");
-
-        // Spawn thread B — it must get LOCK_BUSY, not silently win the lock
-        // by reaping what looks like a stale file.
-        let result = std::thread::spawn(move || ctx_b.lock_workspace())
-            .join()
-            .expect("thread must not panic");
-
-        assert!(result.is_err(), "thread B must not acquire the held lock");
-        let err_msg = result.unwrap_err().to_string();
-        assert!(
-            err_msg.contains("LOCK_BUSY"),
-            "error must indicate LOCK_BUSY (not a stale-reap win), got: {}",
-            err_msg
-        );
-        // Guard still live — lock file must still exist
-        assert!(
-            ctx_a.locks_dir.join("workspace.lock").exists(),
-            "lock file must still exist after thread B was rejected"
-        );
-        let _ = std::fs::remove_dir_all(&dir);
-    }
-}
+mod tests;
